+++ workflow/controller/workflowpod_test.go
@@ -15,7 +15,7 @@ import (
 	wfv1 "github.com/argoproj/argo/pkg/apis/workflow/v1alpha1"
 	"github.com/argoproj/argo/test/util"
 	"github.com/argoproj/argo/workflow/common"
-	"github.com/argoproj/argo/workflow/config"
+	"github.com/argoproj/argo/config"
 )
 
 func unmarshalTemplate(yamlStr string) *wfv1.Template {
